Use hspec hook to have only one TVar
